requirements.txt
psycopg2-binary>=2.9.9
Faker>=21.0.0
tabulate>=0.9.0
